requirements.txt
Flask>=3.0.0
flask-restx>=1.2.0
Flask-Limiter==3.5.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
isort==5.13.2
loguru==0.7.2
flask-login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
email_validator==2.1.1  # per la validazione delle email nei form